feat: Implement network requests using fetch and improve alt text
Replaced xhr request with fetch API, handled response and errors
using then and catch blocks. Configured the url and port of the
DATABASE_URL to match that of the development server's port.
Enhanced the alt text of the restaurant's image for both
restaurant list and restaurant detail pages. Removed unnecessary
commented code.